Add more test cases and move client fixtures to conftest.py
